refactor: move terraform cli detector to right folder
